Rework VorticityCalculator2D in a new file
